Created a sync handler that pulls the login details #23
