+++ app/driver.js
@@ -2,7 +2,9 @@ import Backbone from 'backbone';
 import Marionette from 'backbone.marionette';
 
 import root from './root';
+
 import {Controller as Default, routes as defaultRoutes} from './default/router';
+
 import {
   Controller as Workout, routes as workoutRoutes
 } from './workouts/router';
