+++ app/workouts/collections/workouts.js
@@ -1,7 +1,10 @@
 import Backbone from 'backbone';
 import Workout from '../models/workout';
 
+import {authSync} from '../../base/models/auth';
+
 export default Backbone.Collection.extend({
   model: Workout,
-  url: '/workouts/'
+  url: '/workouts/',
+  sync: authSync
 });
